Source single entrypoint for environment configuration

+++ configure.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
+set -e
 
 BACKUP_RANDOM=backup-$RANDOM
 
 USER_CUSTOM_SH=${HOME}/.custom.sh
 USER_CUSTOM_SH_BACKUP=${USER_CUSTOM_SH}.$BACKUP_RANDOM
-USER_CUSTOM_SH_TMP=${USER_CUSTOM_SH}.tmp
 
 USER_BASHRC=${HOME}/.bashrc
 USER_BASHRC_BACKUP=${USER_BASHRC}.$BACKUP_RANDOM
@@ -14,8 +14,10 @@ R_MAKEVARS=$HOME/.R/Makevars
 echo "Backing up $USER_BASHRC to $USER_BASHRC_BACKUP"
 cp $USER_BASHRC $USER_BASHRC_BACKUP
 
-echo "Backing up $USER_CUSTOM_SH to $USER_CUSTOM_SH_BACKUP"
-cp $USER_CUSTOM_SH $USER_CUSTOM_SH_BACKUP
+if [[ -f $USER_CUSTOM_SH ]]; then
+    echo "Backing up $USER_CUSTOM_SH to $USER_CUSTOM_SH_BACKUP"
+    cp $USER_CUSTOM_SH $USER_CUSTOM_SH_BACKUP
+fi
 
 echo
 
@@ -24,7 +26,7 @@ while true; do
     YN=${YN:-Y}
     case $YN in
     [Yy]*)
-        REPLACE_CUSTOM_SH=1
+        [[ -f $USER_CUSTOM_SH ]] && rm $USER_CUSTOM_SH
         break
         ;;
     [Nn]*)
@@ -64,26 +66,13 @@ while true; do
     esac
 done
 
-echo
 
 source $(dirname $BASH_SOURCE)/scripts/config.sh
 
-cp templates/custom.sh $USER_CUSTOM_SH_TMP
-
-# Render template using environment variables defined in scripts/config.sh
-sed -i".template.backup" -e "s|{{INCLUDE}}|$INCLUDE|g" $USER_CUSTOM_SH_TMP
-sed -i".template.backup" -e "s|{{BIN}}|$BIN|g" $USER_CUSTOM_SH_TMP
-sed -i".template.backup" -e "s|{{LIB}}|$LIB|g" $USER_CUSTOM_SH_TMP
-sed -i".template.backup" -e "s|{{MKL_LIBRARY_PATH}}|$MKL_LIBRARY_PATH|g" $USER_CUSTOM_SH_TMP
-sed -i".template.backup" -e "s|{{PKGCONFIG}}|$PKGCONFIG|g" $USER_CUSTOM_SH_TMP
-sed -i".template.backup" -e "s|{{SHARE}}|$SHARE|g" $USER_CUSTOM_SH_TMP
-rm ${USER_CUSTOM_SH_TMP}.template.backup
-
-if [ $REPLACE_CUSTOM_SH ]; then
-    mv $USER_CUSTOM_SH_TMP $USER_CUSTOM_SH
-else
-    cat $USER_CUSTOM_SH_BACKUP $USER_CUSTOM_SH_TMP >$USER_CUSTOM_SH
-fi
+echo "
+# Source lin-group environment configuration
+source $CONFIG_FILENAME
+" >> $USER_CUSTOM_SH
 
 echo
 echo "Configuration complete! Log out and back in for changes to take effect."
